fix(hitl2): properly show username when set
