+++ modules/hitlnext/src/views/full/app/components/UserProfile.tsx
@@ -1,54 +1,20 @@
 import { Collapsible, lang } from 'botpress/shared'
 import _ from 'lodash'
-import React, { FC, Fragment, useContext, useEffect, useState } from 'react'
+import React, { FC, useState } from 'react'
 
 import { IUser } from '../../../../types'
 import style from '../../style.scss'
-import { Context } from '../Store'
 
-import { generateUsername, getOrSet } from './../utils'
+import UserName from './UserName'
 
-const UserProfile: FC<IUser> = ({ attributes, id }) => {
-  const { state, dispatch } = useContext(Context)
-
+const UserProfile: FC<IUser> = user => {
   const [expanded, setExpanded] = useState(true)
-  const [defaultUsername, setDefaultUsername] = useState<string>()
-
-  useEffect(() => {
-    const key = id
-    const username = getOrSet(
-      () => {
-        return _.get(state, `defaults.user.${key}.username`)
-      },
-      value => {
-        dispatch({
-          type: 'setDefault',
-          payload: {
-            user: {
-              [key]: {
-                username: value
-              }
-            }
-          }
-        })
-      },
-      generateUsername()
-    )
-
-    setDefaultUsername(username)
-  }, [id])
-
-  const userName = () => {
-    return (!_.isEmpty(attributes) && attributes['fullName']) || state.config.defaultUsername
-      ? defaultUsername
-      : lang.tr('module.hitlnext.user.anonymous')
-  }
 
   return (
     <div>
       <div className={style.profileHeader}>
-        <span className={style.clientName}>{userName()}</span>
-        {attributes?.email && <p>{attributes.email}</p>}
+        <UserName user={user} />
+        {user.attributes?.email && <p>{user.attributes?.email}</p>}
       </div>
       <Collapsible
         opened={expanded}
@@ -56,7 +22,7 @@ const UserProfile: FC<IUser> = ({ attributes, id }) => {
         name={lang.tr('module.hitlnext.user.variables.heading')}
         ownProps={{ transitionDuration: 10 }}
       >
-        {!_.isEmpty(attributes) && (
+        {!_.isEmpty(user.attributes) && (
           <table className={style.table}>
             <thead>
               <tr>
@@ -65,7 +31,7 @@ const UserProfile: FC<IUser> = ({ attributes, id }) => {
               </tr>
             </thead>
             <tbody>
-              {Object.entries(attributes).map((entry, index) => (
+              {Object.entries(user.attributes).map((entry, index) => (
                 <tr key={index}>
                   <td>{entry[0]}</td>
                   <td>{entry[1]}</td>
