+++ src/bp/ui-shared/src/Form/FormFields/Upload.tsx
@@ -69,7 +69,10 @@ const Upload: FC<UploadFieldProps> = props => {
     await props.axios
       .post(props.customPath ? props.customPath : 'media', data, { headers: { 'Content-Type': 'multipart/form-data' } })
       .then(response => {
-        const { url } = response.data
+        let url: string = response.data.url
+        if (url.startsWith('/')) {
+          url = `${window.location.protocol}//${window.location.host}${url}`
+        }
 
         dispatch({ type: 'uploadSuccess', data: { url } })
       })
